Struct range::Range

pub struct Range<I> {
    begin: I,
    length: I,
}

A range of indices

Fields

begin

length

Methods

impl<I: RangeIndex> Range<I>

fn new(begin: I, length: I) -> Range<I>

Create a new range from beginning and length offsets. This could be denoted as [begin, begin + length).

   |-- begin ->|-- length ->|
   |           |            |
<- o - - - - - +============+ - - - ->

fn empty() -> Range<I>

fn begin(&self) -> I

The index offset to the beginning of the range.

   |-- begin ->|
   |           |
<- o - - - - - +============+ - - - ->

fn length(&self) -> I

The index offset from the beginning to the end of the range.

               |-- length ->|
               |            |
<- o - - - - - +============+ - - - ->

fn end(&self) -> I

The index offset to the end of the range.

   |--------- end --------->|
   |                        |
<- o - - - - - +============+ - - - ->

fn contains(&self, i: I) -> bool

true if the index is between the beginning and the end of the range.

       false        true      false
         |           |          |
<- o - - + - - +=====+======+ - + - ->

fn is_empty(&self) -> bool

true if the offset from the beginning to the end of the range is zero.

fn shift_by(&mut self, delta: I)

Shift the entire range by the supplied index delta.

                    |-- delta ->|
                    |           |
<- o - +============+ - - - - - | - - - ->
                                |
<- o - - - - - - - +============+ - - - ->

fn extend_by(&mut self, delta: I)

Extend the end of the range by the supplied index delta.

                    |-- delta ->|
                    |           |
<- o - - - - - +====+ - - - - - | - - - ->
                                |
<- o - - - - - +================+ - - - ->

fn extend_to(&mut self, target: I)

Move the end of the range to the target index.

                              target
                                |
<- o - - - - - +====+ - - - - - | - - - ->
                                |
<- o - - - - - +================+ - - - ->

fn adjust_by(&mut self, begin_delta: I, length_delta: I)

Adjust the beginning offset and the length by the supplied deltas.

fn reset(&mut self, begin: I, length: I)

Set the begin and length values.

fn intersect(&self, other: &Range<I>) -> Range<I>

impl<T: Int, I: RangeIndex<Index=T>> Range<I>

Methods for Ranges with indices based on integer values

fn each_index(&self) -> EachIndex<T, I>

Returns an iterater that increments over [begin, end).